+++ Questions/component/QuestionList/SideBar/SideBar.module.css
@@ -1,0 +1,15 @@
+.helpful, #AAdd-btn {
+  background: none;
+  border: none;
+  padding: 0;
+  text-decoration: underline;
+  cursor: pointer;
+}
+
+.display-on {
+  display: block;
+}
+
+.display-off {
+  display: none;
+}
