+++ Questions/css/Modal.css
@@ -1,0 +1,20 @@
+.modal-back {
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  left: 0;
+  top: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.modal-main {
+  position: absolute;
+  background: white;
+  border: 2px solid darkgrey;
+  border-radius: 5px;
+  width: 400px;
+  height: 500px;
+  top:50%;
+  left:50%;
+  transform: translate(-50%,-50%);
+}
